+++ 1-Api/MyPortfolio.Dtos/ProductDto/UpdateProductDto.cs
@@ -10,7 +10,7 @@ namespace MyPortfolio.Dtos.ProductDto
 	{
         public int ProductID { get; set; }
 
-        public int ProductPrice { get; set; }
+        public Decimal ProductPrice { get; set; }
 		public string ProductName { get; set; }
 		public string ProductDescription { get; set; }
 		public string ProductImage { get; set; }
