+++ 2-FrontEnd/MyPortfolio.UI/Dtos/ProductDto/AddProductDto.cs
@@ -5,7 +5,7 @@
 		public int ProductID { get; set; }
 
 
-		public int ProductPrice { get; set; }
+		public Decimal ProductPrice { get; set; }
 
 		public string ProductName { get; set; }
 		public string ProductDescription { get; set; }
