+++ 2-FrontEnd/MyPortfolio.UI/Dtos/ProductDto/ResultProductDto.cs
@@ -3,7 +3,7 @@
     public class ResultProductDto
     {
         public int ProductID { get; set; }
-        public int ProductPrice { get; set; }
+        public Decimal ProductPrice { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public string ProductImage { get; set; }
