+++ 2-FrontEnd/MyPortfolio.UI/Dtos/ProductDto/UpdateProductDto.cs
@@ -3,7 +3,7 @@
     public class UpdateProductDto
     {
         public int ProductID { get; set; }
-        public int ProductPrice { get; set; }
+        public Decimal ProductPrice { get; set; }
         public string ProductName { get; set; }
         public string ProductDescription { get; set; }
         public string ProductImage { get; set; }
